separate pagination from raw in paginateRawAndEntities
Avoiding a change to the principal pagination object, which should
always be the same.

Adding the additional raw items array as a separate return value.

Modified all tests except those using repository to start a transaction
after connecting to the DB and rolling it back before the connection is
closed to avoid isolation error between the tests, where a certain
amount of items was expected.